fix: generation counter u64→u32 for 32-bit ARM cross-compile compatibility
std.atomic.Value(u64) is unsupported on arm-linux-gnueabihf (32-bit).
u32 provides 4 billion generations which is more than sufficient.

+++ src/agent/root.zig
@@ -241,7 +241,7 @@ pub const Agent = struct {
     /// the start of each turn to detect staleness and trigger a refresh of
     /// tool_specs + tools snapshot.  Replaces the old boolean tool_specs_dirty
     /// flag which was never set to true.
-    cached_registry_generation: u64 = 0,
+    cached_registry_generation: u32 = 0,
     mem: ?Memory,
     bootstrap: ?bootstrap_mod.BootstrapProvider = null,
     session_store: ?memory_mod.SessionStore = null,
